pyproject.toml
[tool.poetry]
name = "telegram-stock-bot"
version = "0.1.0"
description = "Telegram Bot to check various stock market related informations"
authors = ["Daniel Ladischenski"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.3"
beautifulsoup4 = "^4.9.3"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
